fix: do not take the whole app down on a build without the speech module
expo-speech-recognition calls requireNativeModule('ExpoSpeechRecognition')
at the top of its own file, which throws on any binary compiled before the
module was added. expo-router loads every route file to build the route
tree, so that throw was not a missing microphone — it was a red screen at
launch, on a dev client that had been working, naming a component nobody had
opened.

The guards that were in place ran too late: they were inside the component,
and the import had already thrown.

The import now lives in DictateVoice.tsx and is reached from DictateButton
through a require inside a try. A binary that knows the module gets the mic;
one that does not gets the plain text field it had before. Both of those are
fine. An app that will not open is not.

Any native module added from here wants the same treatment unless every
binary that will run the bundle is guaranteed to contain it — noted in the
README beside the feature.
